+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - `Vector3 SwizzleXYZ_XZY(Vector3)`
   - `AsRectTransform(Transform)` and `GetRectTransform(Component)`
   - `GetRandomValue` for `Vector2` and `Vector2Int` ranges
+  - `CancelAndDispose(CancellationTokenSource)`
 - `GetComponentCached` and `TryGetComponentCached` now also extend `Component` instead of just `GameObject`
 
 ### Fixed